# docs/openapi.yml
openapi: 3.1.0
info:
  title: Allama API
  summary: Allama API
  description: Allama is the open source automation platform for enterprise. You can
    operate Allama in headless mode by using the API to create, manage, and run
    workflows.
  termsOfService: https://docs.google.com/document/d/e/2PACX-1vQvDe3SoVAPoQc51MgfGCP71IqFYX_rMVEde8zC4qmBCec5f8PLKQRdxa6tsUABT8gWAR9J-EVs2CrQ/pub
  contact:
    name: Allama Founders
    email: founders@allama.com
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
  version: '1'
paths:
...
  /organization/members:
    get:
      tags:
      - organization
      summary: List Org Members
      operationId: organization-list_org_members
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/OrgMemberRead'
                type: array
                title: Response Organization-List Org Members
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
  /organization/members/{user_id}:
    delete:
      tags:
      - organization
      summary: Delete Org Member
      operationId: organization-delete_org_member
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid4
          title: User Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - organization
      summary: Update Org Member
      operationId: organization-update_org_member
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid4
          title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrgMemberRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /organization/sessions:
    get:
      tags:
      - organization
      summary: List Sessions
      operationId: organization-list_sessions
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/SessionRead'
                type: array
                title: Response Organization-List Sessions
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
  /organization/sessions/{session_id}:
    delete:
      tags:
      - organization
      summary: Delete Session
      operationId: organization-delete_session
      security:
      - APIKeyCookie: []
      - APIKeyHeader: []
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          format: uuid4
          title: Session Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    OrgMemberRead:
      properties:
        user_id:
          type: string
          format: uuid4
          title: User Id
        first_name:
          anyOf:
          - type: string
          - type: 'null'
          title: First Name
        last_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Name
        email:
          type: string
          format: email
          title: Email
        role:
          $ref: '#/components/schemas/UserRole'
        is_active:
          type: boolean
          title: Is Active
        is_superuser:
          type: boolean
          title: Is Superuser
        is_verified:
          type: boolean
          title: Is Verified
        last_login_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Login At
      type: object
      required:
      - user_id
      - first_name
      - last_name
      - email
      - role
      - is_active
      - is_superuser
      - is_verified
      - last_login_at
      title: OrgMemberRead
...
    SessionRead:
      properties:
        id:
          type: string
          format: uuid4
          title: Id
        created_at:
          type: string
          format: date-time
          title: Created At
        user_id:
          type: string
          format: uuid4
          title: User Id
        user_email:
          type: string
          format: email
          title: User Email
      type: object
      required:
      - id
      - created_at
      - user_id
      - user_email
      title: SessionRead
...
    UserRole:
      type: string
      enum:
      - basic
      - admin
      title: UserRole
    UserUpdate:
      properties:
        password:
          anyOf:
          - type: string
          - type: 'null'
          title: Password
        email:
          anyOf:
          - type: string
            format: email
          - type: 'null'
          title: Email
        is_active:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Active
        is_superuser:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Superuser
        is_verified:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Verified
        role:
          anyOf:
          - $ref: '#/components/schemas/UserRole'
          - type: 'null'
        first_name:
          anyOf:
          - type: string
          - type: 'null'
          title: First Name
        last_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Name
        settings:
          anyOf:
          - type: object
          - type: 'null'
          title: Settings
      type: object
      title: UserUpdate
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    APIKeyCookie:
      type: apiKey
      in: cookie
      name: fastapiusersauth
    APIKeyHeader:
      type: apiKey
      in: header
      name: x-allama-service-key
...